% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Data.}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{arrival_time}{customer's arrival times}
\item{service_time}{how long each customer was served for}
...
}
}
\usage{
bank
}
\description{
A dataset containing the arrival times and services times for
100 customers
}
\keyword{datasets}
