% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Creates an R file with template}
\usage{
draft(nm = "draft", path = "drafts", template = "template/template.R")
}
\arguments{
\item{nm}{Name of your awesome work}

\item{path}{Directory name}

\item{template}{Template type. Currently only \code{template.R}}
}
\description{
Creates an R file with template
}
\examples{
\donttest{
draft("My Draft")
unlink("drafts", recursive = TRUE)
}
}
