% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{jop_submission}
\alias{jop_submission}
\title{R Markdown Output Formats}
\usage{
jop_submission(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[rmarkdown]{pdf_document}}.}
}
\value{
An R Markdown output format.
}
\description{
R Markdown Output Formats
}
