% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_handlers.R
\name{get_handler}
\alias{get_handler}
\title{Retrieve a handler by type string. Returns NULL for unregistered types.}
\usage{
get_handler(type)
}
\description{
Retrieve a handler by type string. Returns NULL for unregistered types.
}
\keyword{internal}
