% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{get_session_env}
\alias{get_session_env}
\title{Retrieve the captured session environment}
\usage{
get_session_env()
}
\description{
Used internally by app_server to access the user's objects.
}
\keyword{internal}
