% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_with_appendices.R
\name{knit_with_appendices}
\alias{knit_with_appendices}
\title{Convert R Markdown appendices to LaTeX before Knitting Main Document}
\usage{
knit_with_appendices(input, ...)
}
\arguments{
\item{input}{The path to the main R Markdown file}

\item{...}{(Unused)}
}
\description{
Convert R Markdown appendices to LaTeX before Knitting Main Document
}
