% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the repro inline reporting app}
\usage{
launch_app(env = globalenv(), port = NULL, launch.browser = TRUE)
}
\arguments{
\item{env}{The environment to read objects from. Defaults to globalenv().}

\item{port}{Port to run the app on. Defaults to a random available port.}

\item{launch.browser}{Whether to open a browser automatically.}
}
\description{
Call this from your R session after running your analysis chunks. The app
will have access to all objects in your global environment.
}
