% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Draft queue system}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{arrivals of all customers}

\item{ServiceTimes}{service times for all customers}

\item{NumServers}{number of total servers}
}
\value{
a tibble of results
}
\description{
Draft queue system
}
\examples{
Multiserver(1:3, 4:6, 1)
}
