% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_model.R
\name{params_df_to_list}
\alias{params_df_to_list}
\title{Convert parameters data frame to nested list}
\usage{
params_df_to_list(params)
}
\description{
Convert parameters data frame to nested list
}
\keyword{internal}
