% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute_models.R
\name{precompute_model_params}
\alias{precompute_model_params}
\title{Pre-compute model parameters on app startup}
\usage{
precompute_model_params(env)
}
\description{
Pre-compute model parameters on app startup
}
\keyword{internal}
