% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_handlers.R
\name{register_handler}
\alias{register_handler}
\title{Register a class handler}
\usage{
register_handler(
  type,
  priority = 100L,
  badge,
  detect,
  class_label,
  size_label,
  context = function(input) list(),
  prep,
  assign_result = function(result) result,
  setup_code,
  controls = function(nm, type) NULL,
  render
)
}
\description{
Register a class handler
}
\keyword{internal}
