% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{upload}
\alias{upload}
\title{Upload file to Gist. RUN ON CONSOLE}
\usage{
upload(file)
}
\arguments{
\item{file}{Path to file. If missing, gets current open file.}
}
\description{
Upload file or current open file to Gist. To make it work,
create a PAT using \code{\link[usethis:github-token]{usethis::create_github_token()}} and
\code{\link[usethis:edit]{usethis::edit_r_environ()}}.
}
\seealso{
\href{https://happygitwithr.com/credential-caching.html#credential-caching}{PAT tutorial}
}
