% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_count.R
\name{word_count}
\alias{word_count}
\title{Produce a word count for an R Markdown document}
\usage{
word_count(path)
}
\arguments{
\item{path}{A character vector of lenth one; the path to the file for which
a word count is desired}
}
\value{
A named numeric vector of class "wordcount". The vector separately
lists wordc ounts for: (1) text, including section headings and footnotes;
(2) captions and subcaptions; (3) citations in the text; (4) the
references section; and (5) the total word count of all of the above.
}
\description{
When producing a PDF document via LaTeX (via R Markdown), word count options
are often unsatisfactory. Using command line tools such as pdftototex +
wc can notoriously overcount; for example, it would include page numbers as
well as sometimes axis labels (and axis tick labels) in word counts. Many
publishers also consider tools like texcount to notoriously undercount.
This function attempts to get an accurate word count for an R Markdown
document that will be rendered to PDF via LaTeX. As with any option, it will
be imperfect, but it tends to provide counts between using texcount or the
pdftotext + wc options, so it should be fairly accurate.
}
