% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{formats}
\alias{formats}
\alias{jop_submission}
\title{R Markdown Output Formats}
\usage{
jop_submission(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[rmarkdown]{pdf_document}}.}
}
\value{
An R Markdown output format.
}
\description{
A number of additional \code{yaml} options are made available:
\describe{
    \item{tikz}{If \code{true}, the \code{tikz} LaTeX package is loaded.}
    \item{theorems}{If \code{true}, the \code{amsthm} LaTeX package is loaded,
                    and a number of default theorem environments are created:
                    theorem (theorem number prefaced by "Theorem"),
                    proposition (theorem number prefaced by "Proposition"),
                    lemma (theorem number prefaced by "Lemma"),
                    corollary (theorem number prefaced by "Corollary",
                               numbered with reference to the last
                               \code{proposition}; i.e., if after
                               "Proposition 1", the first \code{corollary}
                               will be "Corollary 1.1"),
                    and thm_corollary (theorem number prefaced by "Corollary",
                                       numbered with reference to the last
                                       \code{theorem}).}
}
}
