% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drafr.R
\name{upload}
\alias{upload}
\title{Upload file to Gist (GitHub)}
\usage{
upload(file)
}
\arguments{
\item{file}{Path to file. If missing, gets current open file.}
}
\description{
Upload file or current open file to Gist
}
